requirements.txt
PySide6
Pillow
tifffile
numpy
